refactor classes according to config
