+++ src/main/java/io/jenkins/plugins/pipeline_elasticsearch_logs/write/direct_es/ElasticSearchWriteAccessDirect.java
@@ -3,9 +3,10 @@ package io.jenkins.plugins.pipeline_elasticsearch_logs.write.direct_es;
 import static com.google.common.collect.Range.closedOpen;
 import static io.jenkins.plugins.pipeline_elasticsearch_logs.Utils.logExceptionAndReraiseWithTruncatedDetails;
 
-import com.cloudbees.plugins.credentials.CredentialsProvider;
+import com.cloudbees.plugins.credentials.common.StandardCertificateCredentials;
+import com.cloudbees.plugins.credentials.common.StandardListBoxModel;
+import com.cloudbees.plugins.credentials.common.StandardUsernameListBoxModel;
 import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials;
-import com.cloudbees.plugins.credentials.matchers.IdMatcher;
 import com.google.common.collect.Range;
 
 import java.io.ByteArrayInputStream;
@@ -14,8 +15,10 @@ import java.io.IOException;
 import java.io.PrintStream;
 import java.io.Serializable;
 import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.security.KeyManagementException;
 import java.security.KeyStore;
@@ -23,7 +26,6 @@ import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.CertificateException;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
 import java.util.logging.Level;
@@ -32,6 +34,9 @@ import java.util.logging.Logger;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 
+import hudson.model.Item;
+import hudson.util.FormValidation;
+import hudson.util.ListBoxModel;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.exception.ExceptionUtils;
 import org.apache.http.HttpHost;
@@ -58,12 +63,11 @@ import org.kohsuke.stapler.DataBoundConstructor;
 
 import hudson.Extension;
 import hudson.security.ACL;
-import io.jenkins.plugins.pipeline_elasticsearch_logs.ElasticSearchConfiguration;
-import io.jenkins.plugins.pipeline_elasticsearch_logs.ElasticSearchGlobalConfiguration;
 import io.jenkins.plugins.pipeline_elasticsearch_logs.SSLHelper;
 import io.jenkins.plugins.pipeline_elasticsearch_logs.write.ElasticSearchWriteAccess;
-import jenkins.model.Jenkins;
 import net.sf.json.JSONObject;
+import org.kohsuke.stapler.DataBoundSetter;
+import org.kohsuke.stapler.QueryParameter;
 
 /**
  * Post data to Elastic Search.
@@ -73,7 +77,14 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
 
     private static final Range<Integer> SUCCESS_CODES = closedOpen(200, 300);
 
-    private final URI uri;
+    private String url;
+
+    private URI uri;
+
+    private String certificateId;
+
+    @CheckForNull
+    private String credentialsId;
 
     @CheckForNull
     private String username;
@@ -81,7 +92,7 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
     @CheckForNull
     private String password;
 
-    private final int connectionTimeout;
+    private int connectionTimeout;
 
     @CheckForNull
     private byte[] trustStoreBytes;
@@ -94,26 +105,58 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
 
 	private transient HttpClientContext context;
 
+
+    static final int CONNECTION_TIMEOUT_DEFAULT = 10000;
+
+    public String getUrl() {
+        return url;
+    }
+
+
+    @Nonnull
+    public static Integer ensureCorrectConnectionTimeoutMillis(@CheckForNull Integer value) {
+        if (value == null || value < 0) {
+            return CONNECTION_TIMEOUT_DEFAULT;
+        }
+        return value;
+    }
+
+    @Nonnull
+    public Integer getConnectionTimeoutMillis() {
+        //Ensure correct value - just in case something went wrong
+        return ensureCorrectConnectionTimeoutMillis(connectionTimeoutMillis);
+    }
+
+    @DataBoundSetter
+    public void setConnectionTimeoutMillis(@CheckForNull Integer connectionTimeoutMillis) {
+        this.connectionTimeoutMillis = ensureCorrectConnectionTimeoutMillis(connectionTimeoutMillis);
+    }
+
+    @CheckForNull
+    private Integer connectionTimeoutMillis;
+
     @DataBoundConstructor
     public ElasticSearchWriteAccessDirect() throws URISyntaxException {
-        ElasticSearchConfiguration config = ElasticSearchGlobalConfiguration.get().getElasticSearch();
-        if(config != null) {
-            String credentialsId = config.getCredentialsId();
-            if(credentialsId != null) {
-                StandardUsernamePasswordCredentials credentials = getCredentials(credentialsId);
-                if(credentials != null) {
-                    this.username = credentials.getUsername();
-                    this.password = credentials.getPassword().getPlainText();
-                }
-            }
-            this.uri = new URI(config.getUrl());
-            this.connectionTimeout = config.getConnectionTimeoutMillis();
-        } else {
-            this.username = null;
-            this.password = null;
-            this.uri = null;
-            this.connectionTimeout = -1;
-        }
+    }
+
+
+    public String getCertificateId() {
+        return certificateId;
+    }
+
+    @DataBoundSetter
+    public void setCertificateId(String certificateId) {
+        this.certificateId = certificateId;
+    }
+
+    @CheckForNull
+    public String getCredentialsId() {
+        return credentialsId;
+    }
+
+    @DataBoundSetter
+    public void setCredentialsId(String credentialsId) {
+        this.credentialsId = credentialsId;
     }
 
     /*
@@ -123,7 +166,7 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
         this.username = user;
         this.password = password;
         this.uri = uri;
-        this.connectionTimeout = connectionTimeout;
+        setConnectionTimeoutMillis(connectionTimeout);
         this.trustStoreBytes = trustStoreBytes == null ? null : trustStoreBytes.clone();
     }
 
@@ -146,21 +189,51 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
         public String getDisplayName() {
             return "Direct Elasticsearch Writer";
         }
+
+
+        public static ListBoxModel doFillCredentialsIdItems(@QueryParameter String credentialsId) {
+            StandardUsernameListBoxModel model = new StandardUsernameListBoxModel();
+
+            model.includeEmptyValue().includeAs(ACL.SYSTEM, (Item) null, StandardUsernamePasswordCredentials.class, Collections.emptyList())
+                .includeCurrentValue(credentialsId);
+
+            return model;
+        }
+
+        public static ListBoxModel doFillCertificateIdItems(@QueryParameter String certificateId) {
+            StandardListBoxModel model = new StandardListBoxModel();
+            model.includeEmptyValue().includeAs(ACL.SYSTEM, (Item) null, StandardCertificateCredentials.class)
+                .includeCurrentValue(certificateId);
+
+            return model;
+        }
+
+        public FormValidation doCheckConnectionTimeoutMillis(@QueryParameter("value") Integer value) {
+            int newValue = ensureCorrectConnectionTimeoutMillis(value);
+            if (value == null || newValue != value) {
+                return FormValidation.warning("Illegal value - default " + newValue + " is used instead.");
+            }
+            if (value == 0) return FormValidation.ok("This means no connection timeout is used");
+            return FormValidation.ok();
+        }
+
+        public FormValidation doCheckUrl(@QueryParameter("value") String value) {
+            if (StringUtils.isBlank(value)) {
+                return FormValidation.warning("URL must not be empty");
+            }
+            try {
+                URL url = new URL(value);
+                if (!url.getProtocol().equals("http") && !url.getProtocol().equals("https")) {
+                    return FormValidation.error("Only http/https are allowed protocols");
+                }
+
+            } catch (MalformedURLException e) {
+                return FormValidation.error("URL is not well formed");
+            }
+            return FormValidation.ok();
+        }
     }
 
-    @CheckForNull
-    private static StandardUsernamePasswordCredentials getCredentials(@Nonnull String id) {
-        StandardUsernamePasswordCredentials credential = null;
-        List<StandardUsernamePasswordCredentials> credentials = CredentialsProvider
-                .lookupCredentials(StandardUsernamePasswordCredentials.class, Jenkins.get(), ACL.SYSTEM, Collections.emptyList());
-        IdMatcher matcher = new IdMatcher(id);
-        for (StandardUsernamePasswordCredentials c : credentials) {
-            if (matcher.matches(c)) {
-                credential = c;
-            }
-        }
-        return credential;
-    }
 
     private HttpPost getHttpPost(String data) {
         HttpPost postRequest = new HttpPost(uri);
@@ -233,32 +306,34 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
         return "";
     }
 
-    /**
-     * Posts the given Map to elastic search.
-     *
-     * @param data
-     *            The data to post
-     * @throws IOException
-     */
-    @Override
-    public void push(Map<String, Object> data) throws IOException {
-        String dataString = JSONObject.fromObject(data).toString();
-        HttpPost post = getHttpPost(dataString);
 
-        CloseableHttpResponse response = null;
-        try {
-            CloseableHttpClient httpClient = getClient();
-            response = httpClient.execute(post, context);
-            if (!SUCCESS_CODES.contains(response.getStatusLine().getStatusCode())) {
-                String errorMessage = this.getErrorMessage(response);
-                throw new IOException(errorMessage);
+
+        /**
+         * Posts the given Map to elastic search.
+         *
+         * @param data The data to post
+         * @throws IOException
+         */
+        @Override
+        public void push(Map<String, Object> data) throws IOException {
+            String dataString = JSONObject.fromObject(data).toString();
+            //HttpPost post = getHttpPost(dataString);
+
+            CloseableHttpResponse response = null;
+            try {
+                //     CloseableHttpClient httpClient = getClient();
+                //    response = httpClient.execute(post, context);
+                if (!SUCCESS_CODES.contains(response.getStatusLine().getStatusCode())) {
+                    String errorMessage = this.getErrorMessage(response);
+                    throw new IOException(errorMessage);
+                }
+            } catch (Exception e) {
+                logExceptionAndReraiseWithTruncatedDetails(LOGGER, Level.SEVERE, "Could not push log to Elasticsearch", e);
+            } finally {
+                if (response != null) EntityUtils.consumeQuietly(response.getEntity());
             }
-        } catch (Exception e) {
-            logExceptionAndReraiseWithTruncatedDetails(LOGGER, Level.SEVERE, "Could not push log to Elasticsearch", e);
-        } finally {
-            if(response != null) EntityUtils.consumeQuietly(response.getEntity());
         }
-    }
+
 
     private String getErrorMessage(CloseableHttpResponse response) {
         ByteArrayOutputStream byteStream = null;
@@ -271,7 +346,7 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
                 stream.print("HTTP error code: ");
                 stream.println(response.getStatusLine().getStatusCode());
                 stream.print("URI: ");
-                stream.println(uri.toString());
+           //     stream.println(uri.toString());
                 stream.println("RESPONSE: " + response.toString());
                 response.getEntity().writeTo(stream);
             } catch (IOException e) {
@@ -291,7 +366,7 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
 
     private static class MeSupplier implements Supplier<ElasticSearchWriteAccess>, Serializable {
 
-        private final URI uri;
+        private URI uri;
 
         @CheckForNull
         private String username;
@@ -299,31 +374,17 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
         @CheckForNull
         private String password;
 
-        private final int connectionTimeout;
+        private int connectionTimeout;
 
         @CheckForNull
         private byte[] trustStoreBytes;
 
-        private MeSupplier(ElasticSearchWriteAccessDirect me) throws IOException {
-            ElasticSearchConfiguration config = ElasticSearchGlobalConfiguration.get().getElasticSearch();
-            if (config != null) {
-                String credentialsId = config.getCredentialsId();
-                if (credentialsId != null) {
-                    StandardUsernamePasswordCredentials credentials = getCredentials(credentialsId);
-                    if (credentials != null) {
-                        this.username = credentials.getUsername();
-                        this.password = credentials.getPassword().getPlainText();
-                    }
-                }
-                this.uri = URI.create(config.getUrl());
-                this.connectionTimeout = config.getConnectionTimeoutMillis();
-                this.trustStoreBytes = config.getKeyStoreBytes();
-            } else {
-                this.username = null;
-                this.password = null;
-                this.uri = null;
-                this.connectionTimeout = -1;
-            }
+        private MeSupplier(ElasticSearchWriteAccessDirect me){
+            this.uri = me.uri;
+            this.username = me.username;
+            this.password = me.password;
+            this.connectionTimeout = me.connectionTimeout;
+            this.trustStoreBytes = me.trustStoreBytes;
         }
 
         @Override
@@ -337,9 +398,11 @@ public class ElasticSearchWriteAccessDirect extends ElasticSearchWriteAccess {
     }
 
     @Override
-    public Supplier<ElasticSearchWriteAccess> getSupplier() throws IOException {
+    public Supplier<ElasticSearchWriteAccess> getSupplier() {
         return new MeSupplier(this);
     }
+
+
 
     @Override
     public void close() throws IOException {
